+++ CookieShop.App/DTOs/Order/OrderItemResponse.cs
@@ -3,7 +3,7 @@ namespace CookieShop.App.DTOs.Order;
 public record OrderItemResponse(
     Guid ProductId,
     string ProductName,
-    long UnitPrice,
+    decimal UnitPrice,
     int Quantity,
-    long LineTotal
+    decimal LineTotal
 );
